refactor(openwebif): split the 2096-line client.go by concern
Pure same-package file reorganization (no logic/behavior change): the OpenWebIF
Client god-file is decomposed into cohesive siblings.

- client.go (2096 -> 569): Client/Options/Cacher types, DTOs, New, normalizeOptions
- transport.go: get/doGet, retry/backoff, error classification + wrapping, metrics
- timers.go:    timer CRUD + the A/B-flavor fallback (AddTimer/UpdateTimer/Detect)
- services.go:  bouquets/services + services-schema capability negotiation
- epg.go:       EPG fetch + Latin-1 conversion
- streamurl.go: stream-URL building (WebIF / direct-TS / override)
- status.go:    About/Status/Current/Signal

No behavior change. Build + go vet clean; openwebif tests green; golangci-lint
--new-from-rev = 0; verify-config green.

(Note: the audit's "20+ getJSON copies" was overstated — only ~5 decode sites with
divergent log Msgs, so a getJSON[T] helper would not be byte-identical; the
behavior-preserving win is this decomposition.)
